🐛 Change float type to int

+++ src/home/application/service.py
@@ -22,7 +22,7 @@ class HomeService:
         session: boto3.Session,
         scenario_id: str | None,
         calculate_type: str,
-        percentile: float | None,
+        percentile: int | None,
     ):
         pax_df = await self.home_repo.download_from_s3(session, scenario_id)
         calculator = HomeCalculator(pax_df, calculate_type, percentile)
@@ -43,7 +43,7 @@ class HomeService:
         session: boto3.Session,
         scenario_id: str | None,
         calculate_type: str,
-        percentile: float | None,
+        percentile: int | None,
     ):
         pax_df = await self.home_repo.download_from_s3(session, scenario_id)
         calculator = HomeCalculator(pax_df, calculate_type, percentile)
